prune yt.player.Application instead of .getVideoData
The .getVideoData method has been removed, so we need to parse the
yt.player.Application.create object directly. However this is a large object and
"cyclic object value" have to be dealt with. prune function comes courtesy of:

http://github.com/Canop/JSON.prune

also we need to increase the depth from the default 6 to 9 to catch the video
URLs. Finally I added "use strict" for good measure.

Fix #19